Add ImageClassifier model
Copied without changes from the old repository.

Since the current package of @tensorflow-model/mobilenet on npm depends on tfjs 3.1.0, an overrides section in the package.json was necessary to get this to install against our more recent version. It appears to run with tfjs 4.2.0, but the output of the example sketch seems wrong (see next commit). To be investigated further.